wip: add dashboard, fix max width, truncate bios
Adds a WIP dashboard view for organizations (right now, it just shows an
error message if a user tries to access it via a personal account).

Fixes the max width by setting `box-sizing` to `border-box` and adding
`24px` of padding to the left and right.

Truncates the bios (and other multi-line pieces of text) with ellipsis
using the `dotdotdot-js` package (note that we're still waiting on
FrDH/dotdotdot-js#181 to pass our Typescript build).